refactor(export): Export button moved to navbar and appbar
Closes #796
