+++ packages/geoview-core/src/core/utils/utilities.ts
@@ -185,3 +185,67 @@ export function sanitizeHtmlContent(contentHtml: string) {
   console.log(clean);
   return clean;
 }
+
+/**
+ * Export the map as a PNG
+ * @param {string} mapId Id of map to export
+ */
+export function exportPNG(mapId: string): void {
+  document.body.style.cursor = 'progress';
+  const { map } = api.map(mapId);
+
+  map.once('rendercomplete', () => {
+    const mapCanvas = document.createElement('canvas');
+    const size = map.getSize();
+    // eslint-disable-next-line prefer-destructuring
+    mapCanvas.width = size![0];
+    // eslint-disable-next-line prefer-destructuring
+    mapCanvas.height = size![1];
+    const mapContext = mapCanvas.getContext('2d');
+    Array.prototype.forEach.call(map.getViewport().querySelectorAll('.ol-layer canvas, canvas.ol-layer'), (canvas) => {
+      if (canvas.width > 0) {
+        const opacity = canvas.parentNode.style.opacity || canvas.style.opacity;
+        mapContext!.globalAlpha = opacity === '' ? 1 : Number(opacity);
+        let matrix;
+        const { transform } = canvas.style;
+
+        if (transform) {
+          // Get the transform parameters from the style's transform matrix
+          matrix = transform
+            .match(/^matrix\(([^(]*)\)$/)[1]
+            .split(',')
+            .map(Number);
+        } else {
+          matrix = [parseFloat(canvas.style.width) / canvas.width, 0, 0, parseFloat(canvas.style.height) / canvas.height, 0, 0];
+        }
+
+        // Apply the transform to the export map context
+        CanvasRenderingContext2D.prototype.setTransform.apply(mapContext, matrix);
+        const { backgroundColor } = canvas.parentNode.style;
+        if (backgroundColor) {
+          mapContext!.fillStyle = backgroundColor;
+          mapContext!.fillRect(0, 0, canvas.width, canvas.height);
+        }
+
+        mapContext!.drawImage(canvas, 0, 0);
+      }
+    });
+
+    mapContext!.globalAlpha = 1;
+    mapContext!.setTransform(1, 0, 0, 1, 0, 0);
+
+    try {
+      const image = mapCanvas.toDataURL('image/png');
+      const element = document.createElement('a');
+      const filename = `${mapId}.png`;
+      element.setAttribute('href', image);
+      element.setAttribute('download', filename);
+      element.click();
+    } catch (err) {
+      // eslint-disable-next-line no-console
+      console.error(`Error: ${err}`);
+    }
+  });
+  document.body.style.cursor = 'auto';
+  map.renderSync();
+}
